+++ libera_utils/__init__.py
@@ -1,5 +1,3 @@
 """
 libera_utils
 """
-# Local
-from .db.models import *